fix: bookmark persistence with debug logging
- Fix shortcut type mismatch (u8 -> i32) for JSON compatibility
- Add comprehensive debug logging in Rust backend
- Add console logging in frontend config store
- Log bookmark save/load operations for debugging

+++ src-tauri/src/fs/mod.rs
@@ -27,7 +27,7 @@ pub struct WslDistro {
 pub struct Bookmark {
     pub name: String,
     pub path: String,
-    pub shortcut: Option<u8>, // 1-9 for Ctrl+1..9
+    pub shortcut: Option<i32>,
 }
 
 #[derive(Debug, Serialize, Deserialize, Clone, Default)]
